Still fixing errors with setting defaults...

+++ src/pypetal-jav/utils.py
@@ -38,7 +38,7 @@ def fix_jav_params_after_ufj(javelin_params, drw_rej_res):
 
         if javelin_params['fixed'] is None:
             fixed_tot = np.ones(2 + (nlc-1)*ntophat, dtype=int)
-            p_fix_tot = np.ones(2 + (nlc-1)*ntophat, dtype=int)
+            p_fix_tot = np.ones(2 + (nlc-1)*ntophat )
         else:
             fixed_tot = javelin_params['fixed']
             p_fix_tot = javelin_params['p_fix']
@@ -53,7 +53,7 @@ def fix_jav_params_after_ufj(javelin_params, drw_rej_res):
 
     else:        
         fixed_tot = np.ones( ( nlc-1, 2 + ntophat ), dtype=int )
-        p_fix_tot = np.ones( ( nlc-1, 2 + ntophat ), dtype=int )
+        p_fix_tot = np.ones( ( nlc-1, 2 + ntophat ) )
 
         if javelin_params['fixed'] is not None:
             for i in range(nlc-1):
